adding clarifai api to backend and code review
